Move important parts of 'FollowList' to sub component 'ProfileList'
